done comments controller
