+++ src/components/popup.js
@@ -96,12 +96,7 @@ export default class Popup extends AbstractComponent {
           <div class="form-details__bottom-container">
             <section class="film-details__comments-wrap">
               <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${this._numberComments}</span></h3>
-      
-              <!--<ul class="film-details__comments-list">-->
-                <!---->
-  <!---->
-              <!--</ul>-->
-      
+           
               <div class="film-details__new-comment">
                 <div for="add-emoji" class="film-details__add-emoji-label"></div>
